+++ src/Component/AutoMapper/Tests/Fixtures/Pet.php
@@ -7,10 +7,11 @@ use Symfony\Component\Serializer\Annotation\DiscriminatorMap;
 /**
  * @DiscriminatorMap(typeProperty="type", mapping={
  *    "cat"="Jane\Component\AutoMapper\Tests\Fixtures\Cat",
- *    "dog"="Jane\Component\AutoMapper\Tests\Fixtures\Dog"
+ *    "dog"="Jane\Component\AutoMapper\Tests\Fixtures\Dog",
+ *    "fish"="Jane\Component\AutoMapper\Tests\Fixtures\Fish"
  * })
  */
-class Pet
+abstract class Pet
 {
     /** @var string */
     public $type;